+++ src/hooks/common/index.ts
@@ -1,5 +1,6 @@
 export * from './useDebounce';
 export * from './useCart';
+export { useAddCart } from './useAddCart';
 export * from './useProductListVirtualizer';
 export * from './useProductReviews';
 export * from './useProductOptions';
